GnomeShellWindowManager: Retrieve mouse pos from Mutter
Add a new `GetMousePosition` method to the
DBus interface in Accerciser's GNOME Shell
extension that returns the x and y coords of the
mouse pointer.
GJS API to get the pointer: [1]

Override the base class implementation
`WindowManager#getMousePosition` in the
`GnomeShellWindowManager` subclass to
call that new method.

This makes the "Inspect under mouse" feature
work as expected in a quick test when hovering the
mouse over different UI elements in gedit in
a GNOME Wayland session on Fedora RawHide.

[1]  https://gjs-docs.gnome.org/shell14~14/shell.global#method-get_pointer

Issue: #41