interface_view: Add help text
Show help text in the interface view. That property
was added in at-spi2-core commit [1]

    commit 8f1c70155b4132de2c567b3034e4f242e0450df5
    Author: Mike Gorse <mgorse@suse.com>
    Date:   Wed Jan 3 16:52:31 2024 -0600

        Add HelpText property and corresponding functions in atk and libatspi.

        Closes #146

Currently, at least ATK doesn't notify about changes of
the help text, so updating it in the interface view
automatically when it changes while the object is selected
is not possible as of now.

(I plan to submit a MR to at-spi2-core to implement
sending such a notification in ATK and will submit
another change to Accerciser to handle the update
if that gets accepted.)

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/8f1c70155b4132de2c567b3034e4f242e0450df5