KWin: Retrieve app/window icon via desktop entry
Override the Wnck-specific implementation of
`WindowManager#getApplicationIcon` for
`KWinWindowManager` to retrieve an app
icon from a desktop entry, using the
app ID/desktop file name set for the
top-level windows of the application.

The desktop file name set for a window
is available via the KWin scripting API [1],
so extend the KWin scripts to return that
for each window as well.

Extend `WindowInfo` by adding a new `app_id`
member and set to the desktop file name for the
KWin case.

Add a new `WindowManager#getApplicationIconViaAppID`
helper method that implements the logic to
retrieve an icon, using the desktop file name/app ID
set for the window to retrieve an icon from the
corresponding desktop file according to the
FreeDesktop Desktop Entry Specification [2].

Call that helper method in the new
`KWinWindowManager#getApplicationIcon`
override.

This e.g. makes icon for the Qt-based LibreOffice
variant, Firefox run as a native Wayland client
(GDK_BACKEND=wayland) or VSCodium installed from
FlatHub show up in Accerciser when run on
KDE Plasma Wayland when using Accerciser's KWin
window manager implementation (enabled via
`ACCERCISER_WINDOW_MANAGER=kwin` env var).
Previously, no icon was shown for any of
the above-mentioned apps when they were running
as native Wayland clients.

This adds a new dependency on python3-xdg.

A similar approach will be implemented for
GNOME Shell/Mutter in a separate commit.

[1] https://develop.kde.org/docs/plasma/kwin/api/#kwinwindow
[2] https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html

Issue: #41