GNOME: Use gtk and sandboxed app ID to retrieve app icon
Similar to how the previous commit implemented it for
KWin, override `WindowManager#getApplicationIcon` for
the GNOME Shell/Mutter implementation as well and
reuse the helper method to retrieve an icon via
a desktop file introduced in the previous commit.

Try two kinds of application IDs that are available
for windows via the GNOME Shell extension API, the "gtk application ID"
and the "sandboxed app id". [1]

With this in place, icons are now shown for e.g.
Gedit (which has a gtk application ID of "org.gnome.gedit" set)
and VSCodium installed from Flathub (which has a sandboxed app ID
of "com.vscodium.codium" set) in Accerciser when run
in a GNOME 46 Wayland session on Wayland and using Accerciser's
GNOME Shell WindowManager implementation
(`ACCERCISER_WINDOW_MANAGER=gnomeshell`).

[1] https://gjs-docs.gnome.org/meta14~14/meta.window

Issue: #41