interface_view: Show infos for Hyperlink iface
Add information about the AT-SPI Hyperlink
interface to the corresponding section
in the interface view. Such a section was
there already, but not actually showing
anything except for the section label.

(A part of the information was indirectly
available in the tree view for the Hyperlink
interface of the object that the hyperlink
belongs to, but not when the hyperlink itself
was selected.)

Extend the UI to show the information
specified in the Hyperlink AT-SPI XML
description [1] and implement updating
that information in the existing
`_SectionHyperlink` class that was
only there as a placeholder so far.

A simple example where this gets used is when
the link object in a web page is selected,
e.g. in in Firefox or Chromium with a simple
HTML file like this:

    <html>
    <body>
    <p>
        This is some paragraph with a <a href="https://www.example.org/"> link </a>.
    </p>
    <p>
        This is another paragraph, without any link.
    </p>

    </body>
    </html>

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/2e2b7eb5c4f25e2ee5970e9d4a7e16124cbb2772/xml/Hyperlink.xml