interface view: Show expanders in "Relations" tree view
This makes clearer that the target objects are
children/associated with the nodes showing the
relation type, e.g. that an object is the target
of a "flows to" relation.

Previously, both, the
"<enum atspi flows to of type atspi.relationtype>"
node and the node for the target object looked like
they were both (unrelated) top-level items.
Now, the object is clearly shown as a child.