interface view: Disable "Show" button unless relation target selected
Only enable the "Show" button that can be used to jump
to the currently selected relation target if a relation
target object is actually selected, not when the
node only displaying the type of the relation is
selected, e.g. one displayed as
"<enum atspi flows to of type atspi.relationtype>".

The `_onRelationShow` callback does nothing in case
the relation target type is selected. Disable the button
altogether to make clear that pressing it doesn't have
any effect in that case.