interface view: Don't explicitly set relations view color
No longer use GtkCellRenderer's deprecated "cell-background-gdk"
property [1] to explicitly set a background color for the
"Relations" tree view in the interface view.

Since

    commit 664f2c7c08b2d61947f0e2fac90a47caffee3bb7
    Date:   Fri May 17 15:07:04 2019 +0200

        Fix background color of relations

, the color for both, headers (displaying the relation type)
and objects is the same. (When reverting the above commit for
testing, the same color (black in my setup) is used for both,
font and background for the target object nodes, making the text
for those unreadable.)

I see no visual difference with this commit in place, i.e.
when using the default background rendering logic instead of
explicitly setting the color retrieved via the deprecated
`gtk_style_context_get_background_color` [2].
Therefore, drop explicitly setting a background color altogether.

Presumably, this was used earlier to be able to differentiate
between the nodes showing relation types (headers) and the objects.

This is now solved differently: With previous commit

    commit 05700a03ae3fd693c773c0878691bdeed931c8aa
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Sep 11 08:52:04 2024 +0200

        interface view: Show expanders in "Relations" tree view

in this MR's commit series, the objects are now
distinguishable as they're displayed as children
of the relation type nodes.

This also fixes this warning, seen in the IPython console
after typing something there with an object that has relations
selected in the tree view:

    /usr/lib/python3/dist-packages/pyatspi/registry.py:173: Warning: unable to set property 'cell-background-gdk' of type 'GdkColor' from value of type 'gchararray'
      Atspi.event_main()

[1] https://docs.gtk.org/gtk3/property.CellRenderer.cell-background-gdk.html
[2] https://docs.gtk.org/gtk3/method.StyleContext.get_background_color.html