ipython: Fix autocompletion with IPython 8.29
As the documentation for
`IPython.core.completer.IPCompleter.complete` [1]
describes, the returned tuple contains the
text that was used in the completion
and a list of completion matches for that
text.

While the text returned as used in the completion
always seems to be the full text in IPython 8.20
(at least seen with package python3-ipython 8.20.0-1
on Debian testing), that's not necessarily the
case for newer IPython versions.

At least with python3-ipython 8.29.0-1 on Debian
testing, with an accessible object selected in
Accerciser's treeview, only ".n" would be used
for the matching in the following example:

    In [1]: acc.n
    .name
    .newv
    .notify
    .notify_by_pspec
    In [1]: .n

As a consequence, the text not used for matching
would get lost after pressing the Tab key, and
would have to manually be typed again.

Downgrading to 8.20.0-1 would give the expected
result again:

    In [1]: acc.n
    acc.name
    acc.newv
    acc.notify
    acc.notify_by_pspec
    In [1]: acc.n

Adjust `IterableIPShell.complete` to not rely
on the whole line being used for matching to
make this work as expected with the newer IPython
version as well:
Determine the start index of the text actually
used for completion, and include the text up to
that index in the returned tuple as well (for
both, the completed text and the suggestions
list).

[1] https://ipython.readthedocs.io/en/stable/api/generated/IPython.core.completer.html#IPython.core.completer.IPCompleter.complete