node: Port from deprecated gdk_cairo_create
gdk_cairo_create is deprecated since GDK 3.22 [1].

The documentation says:

> If you are drawing on a native GdkWindow in response to a GDK_EXPOSE
> event you should use gdk_window_begin_draw_frame() and
> gdk_drawing_context_get_cairo_context() instead. GTK will automatically
> do this for you when drawing a widget.

And the gdk_window_begin_draw_frame doc [2] says:

> When using GTK+, the widget system automatically places calls to
> gdk_window_begin_draw_frame() and gdk_window_end_draw_frame() around
> emissions of the GtkWidget::draw signal. That is, if you’re drawing the
> contents of the widget yourself, you can assume that the widget has a
> cleared background, is already set as the clip region, and already has a
> backing store. Therefore in most cases, application code in GTK does not
> need to call gdk_window_begin_draw_frame() explicitly.

In fact, the callback for GtkWidget's `draw` signal [3]
gets the the cairo context to draw to passed as a
parameter, so simply use that instead of creating
a custom one.

Rename _HighLight._onExpose to _HighLight._onDraw
as this is the handler for the "draw" GtkWidget signal.

I see no change in behavior with this commit, i.e.
the highlighting when clicking on an item in the
tree view looks the same.

[1] https://docs.gtk.org/gdk3/func.cairo_create.html
[2] https://docs.gtk.org/gdk3/method.Window.begin_draw_frame.html
[3] https://docs.gtk.org/gtk3/signal.Widget.draw.html