treeview: Ignore event about own app added to AT-SPI registry
In `AccessibleTreeView._accEventChildChanged`, don't only
ignore events where the event source is part of the
current Accerciser application instance, but also
events where the event's `detail1` is part of the application.

This makes the method skip processing of the
"object:children-changed:add" event from the
AT-SPI2 registry/desktop when the Accerciser
application gets registered.

Processing this event could result in an incorrect
list of apps being displayed after startup
(e.g. another app being shown twice, Accerciser missing)
because the logic in `AccessibleTreeview._addChild`
depends on the model being populated when it gets
called from `AccessibleTreeView._accEventChildChanged`
but that was not yet the case when the that event is
received, because `AccessibleTreeVieew._refreshTopLevel`
adds an idle to populate the model instead of doing
it right away, and that idle hasn't been processed yet.

Skip processing for that event. The Accerciser app
is already part of the list of apps in the registry
when the idle gets processed and therefore contained
in the list of apps anyway.

Together with the previous commit

    commit 705462878dc146aa6cd9f37948dcb2a15cd5b994
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jan 21 10:48:09 2025 +0100

        Make check for own app more reliable

, this fixes the app list shown after startup
(at least in all of my tests).