interface_view: Fix Application iface handling
While there is an AT-SPI2 Application interface,
its handling in libatspi is different than that
for other interfaces:

While e.g. for the Text interface, there is
`atspi_accessible_get_text_iface` [1] in libatspi
to acccess the interface and pyatspi provides a
corresponding `Atspi.Accessible.queryText` method
and this is equivalent for most other interfaces,
this is not the case for the Application interface.

(There is only `atspi_accessible_get_application` [2],
but that doesn't provide access to the Application iface
of the current object, but returns the top-level application
object that the current object belongs to.)

For the Application interface, the approach in libatspi
is different: Access to the Application properties
is provided via the AtspiAccessible class, and the documentation
states that the corresponding functions only work for
application root objects, e.g. for
`atspi_accessible_get_toolkit_version`[3]:

> Gets the toolkit version for an AtspiAccessible object.
> Only works on application root objects.

Therefore, don't try to call a non-existing method
to get access to the interface, but simply call the
existing Atspi.Accessible methods instead.

Without this commit in place, Accerciser's GUI wouldn't
show up any more when used with a libatspi version
containing commit [4]

    commit 16895d31b0eb9c7985e54a6fc77691b894d7653e
    Author: Lukáš Tyrychtr <ltyrycht@redhat.com>
    Date:   Wed Apr 23 10:24:52 2025 +0200

        atspi-accessible: Report Application interfaces in get_interfaces

        It was forgotten in the checks, so fix that.

        Fixes #192.

that implemented reporting of the AtspiCollection
interface, i.e. this code path would be triggered now
and fail.

While at it, also, switch from the pyatspi-specific
versions of the methods to the plain AtSpi ones.

[1] https://docs.gtk.org/atspi2/method.Accessible.get_text_iface.html
[2] https://docs.gtk.org/atspi2/method.Accessible.get_application.html
[3] https://docs.gtk.org/atspi2/method.Accessible.get_toolkit_version.html
[4] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/16895d31b0eb9c7985e54a6fc77691b894d7653e