interface_view: Show AT-SPI version
Get the AT-SPI (IPC spec) version and display
it in the section for the Application interface
infos, as this is a property of the Application
interface.

As the XML spec says, this should currently always
be "2.1" [1]:

    <!--
        AtspiVersion: You should return "2.1" here.

        This was intended to be the version of the atspi interfaces
        that the application supports, but atspi will probably move to
        using versioned interface names instead.  Just return "2.1" here.
    -->
    <property name="AtspiVersion" type="s" access="read"/>

Handle null/None values gracefully and simply display
an empty string, as this was seen e.g. being returned for
Qt applications like "kded6" in a KDE Plasma session on
Debian testing.
Related pending Qt change to report the correct version: [2]

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/2d83e3a6dfa9e10243b296179a2ba85e37bdf1cc/xml/Application.xml#L19-26
[2] https://codereview.qt-project.org/c/qt/qtbase/+/646896