interface view: Show Accessible's locale
In the interface view's "Accessible" section, add an entry
for the AT-SPI Accessible "Locale" property [1]
that was added in at-spi2-core commit [2]

    commit 746fb37ba66ce815e27080eda8a54101c96e8eec
    Author: Mike Gorse <mgorse@suse.com>
    Date:   Mon Feb 25 18:36:12 2013 -0600

        Add atspi_accessible_get_object_locale

        https://bugzilla.gnome.org/show_bug.cgi?id=694368

This e.g. displays the value "en_GB.UTF-8" when selecting
the accessible object for the Gedit application in Accerciser's
tree view when running a desktop session using that locale.

Related pending Qt change to support that property
in Qt applications: [3]
(Without that change in place, an empty string will
be displayed for any object in Qt applications.)

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/f363576e051bf1363af50a7dd7100045aa5a8075/xml/Accessible.xml#L55-67
[2] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/746fb37ba66ce815e27080eda8a54101c96e8eec
[3] https://codereview.qt-project.org/c/qt/qtbase/+/647665